Updated automatic script

+++ test/compilemu.sh
@@ -3,13 +3,18 @@
 
 infile=$1
 outfile=${infile%.*}.txt
-pdp8exe=~/src/simh_mod/BIN/pdp8
+errfile=error.txt
+pdp8exe=pdp8_music
 time2wav=../wav_creator_buf/time2wav
 
-$pdp8exe pdp8.ini $1 > $outfile
-sed -i '' -e '1,11d' $outfile
-sed -i '' -e '$ d' $outfile
-tr -d '\r' < $outfile > temp.txt
-tr -dc '0-9\n' < temp.txt > $outfile
-rm temp.txt
+if test $infile -nt $outfile ; then
+	$pdp8exe pdp8.ini $1 > $outfile
+	sed -i '' -e '1,11d' $outfile
+	head $outfile > $errfile
+	sed -i '' -e '$ d' $outfile
+	tr -d '\r' < $outfile > temp.txt
+	tr -dc '0-9\n' < temp.txt > $outfile
+	rm temp.txt
+	cat $errfile
+fi
 $time2wav $outfile
